Fix userconfig SQL failure from overflowing JIT capture indices (#874)
* Avoid unrelated DDL locks during cache-trigger cleanup

* Fix numbered-local overflow in wide JIT closures

* Keep outer invocation slots out of JIT local frame sizing

* Capture stalled cache-manager stacks in CI temporarily

* Keep CI test invocation intact while collecting diagnostics

* Publish blob RAM size without waiting for readers during SHOW

* Transfer shard RAM ownership before publishing commit forwarding

+++ scm/scm.go
@@ -933,8 +933,9 @@ func closeProcedureCaptures(expression Scmer, callFrame *Env, bound map[Symbol]s
 	return NewSlice(closed)
 }
 
-// helper pseudo type to optimize parameter reading from indices
-type NthLocalVar uint8 // equals to (var i)
+// Numbered locals include the captures appended when the JIT closes a lambda.
+// Wide projections can exceed 255 slots even when each SQL row is smaller.
+type NthLocalVar uint32 // equals to (var i)
 
 /*
  Environments
